Moved the two files from ui to the base folder
