+++ app/src/main/java/smyo/app/ItemActivity.kt
@@ -1,4 +1,4 @@
-package smyo.app.ui
+package smyo.app
 import kotlinx.android.synthetic.main.add_item.*
 
 class ItemActivity {
